do some corrections
merge some test files
correct a wrong path
