--- docs/shimin.md
@@ -1,1 +1,0 @@
-市民
